+++ KUMA/app/src/main/AndroidManifest.xml
@@ -10,13 +10,14 @@
             android:roundIcon="@mipmap/ic_launcher_round"
             android:supportsRtl="true"
             android:theme="@style/AppTheme">
-        <activity android:name=".TitleController">
+        <activity android:name=".Schedule">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
 
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
+        <activity android:name=".ScheduleOneday"/>
     </application>
 
 </manifest>
