+++ KUMA/app/src/main/res/layout/activity_schedule.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<android.support.constraint.ConstraintLayout
+        xmlns:android="http://schemas.android.com/apk/res/android"
+        xmlns:tools="http://schemas.android.com/tools"
+        xmlns:app="http://schemas.android.com/apk/res-auto"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        tools:context=".Schedule">
+
+    <CalendarView
+            android:layout_height="373dp"
+            android:id="@+id/calendarView"
+            android:layout_alignParentTop="true"
+            android:layout_alignParentLeft="true"
+            android:layout_alignParentStart="true"
+            android:showWeekNumber="false"
+            android:layout_width="410dp"/>
+
+</android.support.constraint.ConstraintLayout>
