+++ KUMA/app/src/main/res/layout/activity_schedule_oneday.xml
@@ -1,0 +1,83 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout
+        xmlns:android="http://schemas.android.com/apk/res/android"
+        xmlns:tools="http://schemas.android.com/tools"
+        xmlns:app="http://schemas.android.com/apk/res-auto"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:orientation="vertical"
+        tools:context=".ScheduleOneday">
+
+    <EditText
+            android:id="@+id/edt_id"
+            android:hint="@string/id"
+            android:inputType="number"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content" android:autofillHints=""/>
+    <EditText
+            android:id="@+id/edt_title"
+            android:hint="@string/title"
+            android:inputType="text"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content" android:autofillHints=""/>
+    <EditText
+            android:id="@+id/edt_timeBegin"
+            android:hint="BEGIN 00:00"
+            android:inputType="time"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content" android:autofillHints=""/>
+    <EditText
+            android:id="@+id/edt_timeEnd"
+            android:hint="END 00:00"
+            android:inputType="time"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content" android:autofillHints=""/>
+    <EditText
+            android:id="@+id/edt_place"
+            android:hint="PLACE"
+            android:inputType="text"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content" android:autofillHints=""/>
+    <EditText
+            android:id="@+id/edt_memo"
+            android:hint="MEMO"
+            android:inputType="text"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content" android:autofillHints=""/>
+
+    <LinearLayout
+            android:orientation="horizontal"
+            android:weightSum="3"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content">
+
+        <Button
+                android:id="@+id/btn_add"
+                android:text="@string/add"
+                android:layout_weight="1"
+                android:layout_width="0dp"
+                android:layout_height="wrap_content"/>
+        <Button
+                android:id="@+id/btn_update"
+                android:text="@string/update"
+                android:layout_weight="1"
+                android:layout_width="0dp"
+                android:layout_height="wrap_content"/>
+        <Button
+                android:id="@+id/btn_delete"
+                android:text="@string/delete"
+                android:layout_weight="1"
+                android:layout_width="0dp"
+                android:layout_height="wrap_content"/>
+
+    </LinearLayout>
+
+    <ListView
+            android:id="@+id/list_plan"
+            android:choiceMode="singleChoice"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content">
+
+    </ListView>
+
+</LinearLayout>
